1. Finish testing PoaNetworkConsensus 2. Rename SimpleConsensus to PoaNetworkConsensus
